requirements.txt
# Solar Geo AI — Python dependencies
# Install: pip install -r requirements.txt
# Python 3.11+

requests==2.31.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
xgboost==2.0.3
matplotlib==3.8.4
seaborn==0.13.2
streamlit==1.33.0
folium==0.16.0
streamlit-folium==0.20.0
groq==0.5.0
python-dotenv==1.0.1
jupyter==1.0.0
ipykernel==6.29.4